Introduced support for playlists!
Implemented a queue management page.
Implemented a queue preview in the main dashboard.
Implemented a chat box in the main dashboard.
Introduced the ability for modules to provide completely custom queuing logic.
Introduced the ability to import songs from a user's favorites collection.
Implemented a variety of helper functions in the different managers for module use.